A very rough migration of my code (originally in a private BitBucket repo) with the personal information scrubbed (server key's, gedcom, and images not being committed). This will of course lose some history, but it's a necessary sacrifice!
When I originally started work on this, I made use of MySQL as a persistence for the data parsed out from the provided gedcom (a plaintext genealogical file format) as well as user data, however my plan is to ditch that and
switch to something different. For this initial commit,

I've started restructing the code a bit (originally it was all under src/jonware/... - not best-practice golang, admittedly!) so there's a good chance it's not quite in a compileable state atm, but I wanted to get this in
asap (especially as I need to replace the MySQL dependency).